+++ packages/fighting-design/tag/src/tag.vue
@@ -1,6 +1,6 @@
 <script lang="ts" setup name="FTag">
   import { Props } from './props'
-  import { ref, reactive } from 'vue'
+  import { ref, reactive, toRefs } from 'vue'
   import { FSvgIcon } from '../../svg-icon'
   import { FCloseBtn } from '../../close-btn'
   import { useList, useRun, useGlobal } from '../../_hooks'
@@ -11,7 +11,7 @@
 
   /** 替换 type 后得到的 props */
   const params = reactive({
-    ...prop,
+    ...toRefs(prop),
     type: getType()
   })